Decouple display and window from `RawWindowHandle`
On free platforms there're usually multiple displays, so APIs like
EGL have a concept of native display when creating EGLDisplay.

Sometimes creating EGLDisplay must be done before creating the window
on that particular display, to supply e.g. X11 visual information to
the window that will be created on that display later on.