Expand documentation of RawDisplayHandle
I came to this crate after reading about the recent changes to
[`glutin`][0]. At a glance I didn't understand the difference between
`RawWindowHandle` and `RawDisplayHandle` (or even what a display was),
and the enums looked mostly the same.

After reading through this [issue][1] and digging through the structs, I
have a better understanding of what each type represents, but I thought
it would be useful to add a line clarifying the distinction between the
two. I also added some examples of what external types may be
represented for a few platforms.

[0]: https://github.com/rust-windowing/glutin/releases/tag/v0.30.0
[1]: https://github.com/rust-windowing/raw-window-handle/issues/92