docs: Clarify Win32 handle validity
This commit clarifies what is expected by the system for a Win32 handle.
It is expected that the Win32 handle belongs to the current thread. This
is not a breaking change as it was previously necessary for the window
to be considered "valid".

cc https://github.com/rust-windowing/winit/pull/3593

Signed-off-by: John Nunley <dev@notgull.net>